split ts files
